outside temp fix and database fix

+++ database/migrations/2023_05_24_095106_create_room_times_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('room_id');
             $table->dateTime('time');
             $table->integer('co2');
-            $table->float('temperature');
+            $table->integer('temperature');
             $table->integer('outside_temperature')->nullable();
             $table->boolean('booked')->nullable();
             $table->timestamps();
